[LAB6] Implement visitor on book elements
The implementation is further explained below:

A generic visitor is required in order to extend
the range of operations that are possible on a book
without having to go into each element type to add
a function.

The TableOfContents class makes use of TocGenerator.
It now expects to be given a reference to the book
so that it can generate the book's ToC on-demand.
The TocGenerator also takes into account the
"width" and "height" of the TOC (how deep in the
book tree is a given element).

The statistics class also works as expected,
being able to print a few counters regarding the
contents of the book.